requirements.txt
anthropic==0.18.1
beautifulsoup4==4.12.2
google-api-python-client==2.97.0
PyPDF2==3.0.1
PyYAML==6.0.1
requests==2.31.0